@ymarco 2023-04-16
> Thank you for taking the time to profile the runtime and for providing the PR.

My pleasure, optimizing code to run faster is one of the things I love best in programming.

> I guess, the same caching mechanism can also be used to speed up the processing of vector fonts as it would reduce the number of many FreeType function calls that can take up a significant amount of execution time. So, rather than adding another caching method, I'd first like to check whether the existing code can be extended to vector fonts and if it reduces the runtime sufficiently.

Sure. I had no idea this other cache existed, hiding under my nose.

> Some further tweaks can also be applied to the formatting output which is usually quite slow when using the STL ostream operators. There are some libraries, like [{fmt}](https://github.com/fmtlib/fmt), that look promising and might help to speed up outputting the SVG code. If I find some time in the next few weeks, I'll have a closer look into this as well.

I never thought about that, might help a little.

The next thing I thought of doing was trying to get down the 131ms conversion time of one fragment, since it's a common case when the user is just finished typing a new fragment. Profiling there shows that a lot of time is spent reading the map file (in my case `/var/lib/texmf/fonts/map/dvips/updmap/ps2pk.map` which has 36k lines):
![image](https://user-images.githubusercontent.com/40410676/232276915-965d6c36-4a55-4ca8-943e-c5c808324e03.png)

 I'll try to get a something working next weekend or the one after that, hopefully