@mgieseki 2023-04-16
> The next thing I thought of doing was trying to get down the 131ms conversion time of one fragment, since it's a common case when the user is just finished typing a new fragment. Profiling there shows that a lot of time is spent reading the map file (in my case `/var/lib/texmf/fonts/map/dvips/updmap/ps2pk.map` which has 36k lines).
> 
> I'll try to get a something working next weekend or the one after that, hopefully

Yes, that's definitely an area with room for improvement. I suppose, one way to speed up reading the mapping information could be to write the collected `FontMap::Entry` data of a `.map` file to a binary file stored in dvisvgm's cache directory. If the time stamps of these files are newer than those of the corresponding `.map` files, the data can be read in from the binary files much faster because there's no need to parse the `.map` files again and again.