@ymarco 2023-04-16
Yeah, I thought about this, but I don't know if parsing it from binary would be much better than parsing the `.map`, which already looks pretty straightforward. We'll see, I guess