@ymarco 2023-04-16
Would be also cool to maybe read the entire map file (or cache) file to memory and mess with [C++17 `string_view`s](https://en.cppreference.com/w/cpp/string/basic_string_view) to avoid allocating all of those `std::string`s. Though we wouldn't be able to close the file or release the memory then, and I don't know what's the consensus about using C++17 stuff.